Updated model - Standardization of features, new json files
This marks a few new items that were added. Primarly, I added two new json files that are going to be very useful to improving the model:

1. song_features_std
The "std" refers to standardized features. The values in this file are nothing but the standardized features. It is worth noting that standardization added a marked improvement in prediction accuracy, so its a feature we are going to keep as we create the database/json files

2. song_features_ext
the "ext" stands for extended. If you open this file, you while quickly see that entries in our "database" are much much larger. The idea here is that I'm going to greatly expand the number of features we pull from database entries, and then feed them into a CNN to try and extract a more robust set of features for the classification task. This works because the MFCCs are a time-series in essence, so this should add more parameters to the model, but improve the performance in turn.

One more thing: because song_features_ext.json is huge, I don't want to push it. Need to talk about this issue later.